+++ src/Eloquent/Enumeration/ValueMultitonInterface.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of the Enumeration package.
+ *
+ * Copyright © 2013 Erin Millard
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Eloquent\Enumeration;
+
+/**
+ * The interface implemented by Java-style enumeration instances with a value.
+ */
+interface ValueMultitonInterface extends MultitonInterface
+{
+    /**
+     * Returns the value of this member.
+     *
+     * @return scalar The value of this member.
+     */
+    public function value();
+}