+++ test/src/Eloquent/Enumeration/Test/Fixture/InvalidValueMultiton.php
@@ -11,20 +11,12 @@
 
 namespace Eloquent\Enumeration\Test\Fixture;
 
-use Eloquent\Enumeration\AbstractMultiton;
-
-abstract class TestMultiton extends AbstractMultiton
+class InvalidValueMultiton extends ValidValueMultiton
 {
     protected static function initializeMembers()
     {
         parent::initializeMembers();
 
-        static::$calls[] = array(
-            get_called_class() . '::' . __FUNCTION__,
-            func_get_args()
-        );
-
-        new static('FOO', 'oof');
-        new static('BAR', 'rab');
+        new static('QUX', 'xuq');
     }
 }