+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 # Enumeration changelog
 
-### 4.0.0 (2013-08-13)
+## 5.0.0 (2013-11-11)
+
+- **[BC BREAK]** Renamed classes:
+    - `Multiton` -> `AbstractMultiton`
+    - `Enumeration` -> `AbstractEnumeration`
+- **[BC BREAK]** Exceptions no longer extend `LogicException`, but instead
+  directly extend from the base `Exception` class.
+- **[NEW]** Implemented `AbstractValueMultiton`, an an abstract base class for
+  implementing multitons with values. `AbstractEnumeration` now extends from
+  this base class.
+- **[NEW]** Implemented `AbstractUndefinedMemeberException`, an abstract base
+  class for implementing custom undefined member exceptions.
+  `UndefinedMemberException` now extends from this base class.
+- **[NEW]** Multiton instances now implement formal interfaces:
+    - `AbstractMultiton` instances implement `MultitonInterface`.
+    - `AbstractValueMultiton` instances implement `ValueMultitonInterface`.
+    - `AbstractEnumeration` instances implement `EnumerationInterface`
+- **[NEW]** Implemented `membersBy()` and `membersByPredicate()` which help to
+  retrieve sets of members by various criteria.
+
+## 4.0.0 (2013-08-13)
 
 - **[BC BREAK]** `Multiton` method name changes:
     - `multitonInstances` -> `members`
@@ -19,7 +39,7 @@
 - **[NEW]** Defaulting variants of member search methods
 - **[MAINTENANCE]** General repository maintenance
 
-### 3.0.2 (2013-03-04)
+## 3.0.2 (2013-03-04)
 
 - **[NEW]** [Archer] integration
 - **[NEW]** Implemented changelog